requirements.txt
aiohttp==3.7.3
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
braceexpand==0.1.6
chardet==3.0.4
discord.py==1.6.0a2705+geb11079
humanize==3.2.0
idna==2.10
import-expression==1.1.4
jishaku==1.20.0.220
multidict==5.1.0
PyYAML==5.3.1
six==1.15.0
style==1.1.0
timeago==1.0.15
typing-extensions==3.7.4.3
update==0.0.1
yarl==1.6.3